+++ frontend/src/app/main/data/workspace.cljs
@@ -253,17 +253,24 @@
        (->> (rx/of ::dwp/finalize)
             (rx/observe-on :async))))))
 
+(declare go-to-page)
 (defn initialize-page
   [page-id]
   (us/assert ::us/uuid page-id)
   (ptk/reify ::initialize-page
+    ptk/WatchEvent
+    (watch [_ state _]
+      (when-not (contains? (get-in state [:workspace-data :pages-index]) page-id)
+        (let [default-page-id (get-in state [:workspace-data :pages 0])]
+          (rx/of (go-to-page default-page-id)))))
+    
     ptk/UpdateEvent
     (update [_ state]
       (let [;; we maintain a cache of page state for user convenience
             ;; with the exception of the selection; when user abandon
             ;; the current page, the selection is lost
-
             page    (get-in state [:workspace-data :pages-index page-id])
+            page-id (:id page)
             local   (-> state
                         (get-in [:workspace-cache page-id] workspace-local-default)
                         (assoc :selected (d/ordered-set)))]